tossPay 결제 및 취소

+++ src/main/java/com/eatpizzaquickly/jariotte/domain/payment/dto/response/PostPaymentResponse.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Getter
 public class PostPaymentResponse {
     private final String payUid;
-    private final int amount;
+    private final Long amount;
     private final String payInfo;
     private final PayMethod payMethod;
     private final PayStatus payStatus;
